Version 2:
When a button is pressed, all buttons disappear and the user is told if they were correct in answering or incorrect.
If the correct answer was selected, a score counter is incremented.
If the score counter is greater than the top score in the SharedPreferences file, write the new score to the SharedPreferences file.
If the answer was incorrect, reset the score counter to 0.
Present a new question to the user